Put layout on _root layout
