--- packages/ui/src/layout.tsx
@@ -1,18 +1,0 @@
-import { Link } from "@tanstack/react-router";
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return (
-    <div className="w-screen h-screen flex flex-col">
-      <div className="p-2 flex gap-2">
-        <Link to="/" className="[&.active]:font-bold">
-          Home
-        </Link>
-        <Link to="/about" className="[&.active]:font-bold">
-          About
-        </Link>
-      </div>
-      <hr />
-      <main className="flex flex-1">{children}</main>
-    </div>
-  );
-}
