--- packages/ui/src/routes/about.tsx
@@ -1,16 +1,0 @@
-import Layout from "@/layout";
-import { createFileRoute } from "@tanstack/react-router";
-
-export const Route = createFileRoute("/about")({
-  component: About,
-});
-
-function About() {
-  return (
-    <Layout>
-      <div className="flex flex-1 items-center justify-center">
-        <h3>About Page</h3>
-      </div>
-    </Layout>
-  );
-}
